requirements.txt
rich>=13.0.0
google-generativeai>=0.5.0
